use VttThumbnailDownloader instead of HttpGetWebVttThumbnails
